+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IndividualAccountRequest resource
 - IndividualAccountRequest::Address sub-resource
 - IndividualAccountAttachment resource
+- data, description, due, expired and jws attributes to BrcodePreview resource
+### Fixed
 
 ## [0.6.0] - 2026-07-16
 ### Added